break up sources for debugger/src/server
