+++ debugger/src/server/session_data.rs
